Cleaned up Main.kt
