+++ .run/MainKt.run.xml
@@ -6,7 +6,7 @@
     <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="false" />
     <option name="ALTERNATIVE_JRE_PATH" />
     <option name="PASS_PARENT_ENVS" value="true" />
-    <option name="MAIN_CLASS_NAME" value="org.kamiblue.botkt.MainKt" />
+    <option name="MAIN_CLASS_NAME" value="org.kamiblue.botkt.Main" />
     <option name="WORKING_DIRECTORY" value="file://$PROJECT_DIR$/run" />
     <method v="2">
       <option name="Make" enabled="true" />
